Rename node_str.rs to name.rs
