+++ crates/apollo-compiler/src/name.rs
